requirements.txt
django==3.1.3
django-stubs==1.8.0
firebase-admin==5.0.1
mypy==0.910
pytest==6.2.4
python-dateutil==2.8.1
pytest-django==4.4.0
pytest-cov==2.12.1
pytest-mypy==0.8.1
types-python-dateutil==0.1.4